+++ frontend/webpack.config.js
@@ -2,7 +2,7 @@
 
 const path = require('path');
 const MiniCssExtractPlugin = require(`mini-css-extract-plugin`);
-const VueLoaderPlugin = require('vue-loader/lib/plugin');
+const { VueLoaderPlugin } = require('vue-loader');
 
 const isDevelopmentMode = process.env.NODE_ENV !== 'production';