+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+2023.08
+-------------------------------------------------------------------------------
+- Added django CMS 3.11 support
+- Added django 4.2 LTS support
+- Python 3.10 support
+- Node 18.17 LTS support
+- Updated almost all backend dependencies
+- Updated almost all frontend dependencies
+- Removed light_gallery plugin
+
+
 2021.08
 -------------------------------------------------------------------------------
 - completely removed the aldryn stack (aldryn-addons, aldryn-django, aldryn-djangocms)